api/requirements_dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements_dev.txt requirements_dev.in
#
argh==0.26.2              # via watchdog
attrs==17.2.0             # via flake8-bugbear
bandit==1.4.0
boto3==1.4.7
botocore==1.7.4
certifi==2017.7.27.1
cfenv==0.5.3
chardet==3.0.4
colorama==0.3.9           # via pytest-watch
coverage==4.4.1           # via pytest-cov
decorator==4.1.2
dj-database-url==0.4.2
django-admin-interface==0.5.2
django-cas-ng==3.5.8
django-colorfield==0.1.14
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-filter==1.0.4
django-flat-responsive==1.4.1
django-flat-theme==1.1.4
django-reversion==2.0.10
django-storages==1.6.5
django==1.11.5
djangorestframework==3.6.4
docopt==0.6.2             # via pytest-watch
docutils==0.13.1
flake8-bugbear==17.4.0
flake8-builtins==1.0
flake8-comprehensions==1.4.1
flake8-isort==2.2.2
flake8-pep3101==1.0
flake8-polyfill==1.0.1    # via flake8-isort
flake8-print==2.0.2
flake8-string-format==0.2.3
flake8-tuple==0.2.13
flake8==3.4.1
furl==1.0.0
gitdb2==2.0.2             # via gitpython
gitpython==2.1.5          # via bandit
gunicorn==19.7.1
idna==2.6
isort==4.2.15             # via flake8-isort
jedi==0.10.2
jmespath==0.9.3
lxml==4.1.0
mccabe==0.6.1             # via flake8
model-mommy==1.3.2
networkx==2.0
newrelic==2.90.0.75
orderedmultidict==0.7.11
pathtools==0.1.2          # via watchdog
pbr==3.1.1                # via stevedore
pdfminer.six==20170720
pep8-naming==0.4.1
psycopg2==2.7.3.1
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via flake8
pycryptodome==3.4.7
pyflakes==1.5.0           # via flake8
pyinotify==0.9.6
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-watch==4.1.0
pytest==3.2.1
python-cas==1.2.0
python-dateutil==2.6.1
pytz==2017.2
pyyaml==3.12              # via bandit, watchdog
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
smmap2==2.0.3             # via gitdb2
sqlparse==0.2.3           # via django-debug-toolbar
stevedore==1.26.0         # via bandit
testfixtures==5.2.0       # via flake8-isort
tqdm==4.19.5
urllib3==1.22
watchdog==0.8.3           # via pytest-watch
whitenoise==3.3.0
